Fixed the pong screen leaving a trail
Fixed the pong screen leaving a trail when the ball and the paddles move around the screen.
This was caused by the screen not being cleared before the ball or paddles are drawn.
This was fixed by adding a call to clear the screen before the ball or paddles are drawn.